--- core/internal/shared/shared.go
@@ -1,2 +1,0 @@
-// Package shared has routines shared between core server and client
-package shared
