+++ core/pkg/server/stream_init.go
@@ -16,11 +16,11 @@ import (
 	"github.com/wandb/wandb/core/internal/filetransfer"
 	"github.com/wandb/wandb/core/internal/runfiles"
 	"github.com/wandb/wandb/core/internal/settings"
-	"github.com/wandb/wandb/core/internal/shared"
 	"github.com/wandb/wandb/core/internal/waiting"
 	"github.com/wandb/wandb/core/internal/watcher2"
 	"github.com/wandb/wandb/core/pkg/filestream"
 	"github.com/wandb/wandb/core/pkg/observability"
+	"github.com/wandb/wandb/core/pkg/utils"
 )
 
 // NewBackend returns a Backend or nil if we're offline.
@@ -92,7 +92,7 @@ func NewFileStream(
 		Settings:  settings.Proto,
 		Logger:    logger,
 		ApiClient: fileStreamRetryClient,
-		ClientId:  shared.ShortID(32),
+		ClientId:  utils.ShortID(32),
 	}
 
 	return filestream.NewFileStream(params)
