+++ core/pkg/utils/headfoot.go
@@ -1,4 +1,4 @@
-package shared
+package utils
 
 import (
 	"fmt"
