+++ core/pkg/utils/shortid_test.go
@@ -1,15 +1,15 @@
-package shared_test
+package utils_test
 
 import (
 	"testing"
 
 	"github.com/stretchr/testify/assert"
-	"github.com/wandb/wandb/core/internal/shared"
+	"github.com/wandb/wandb/core/pkg/utils"
 )
 
 func TestShortID(t *testing.T) {
 	t.Run("shortid", func(t *testing.T) {
-		item := shared.ShortID(8)
+		item := utils.ShortID(8)
 		assert.Equal(t, len(item), 8)
 	})
 }
