+++ experiment/client-go/bindings/core/generate.go
@@ -1,4 +1,4 @@
-package main
+package core
 
 import (
 	_ "embed"
