+++ experiment/client-go/bindings/core/partialdata.go
@@ -1,4 +1,4 @@
-package main
+package core
 
 import (
 	"sync"
