+++ experiment/client-go/bindings/core/runkeeper.go
@@ -1,4 +1,4 @@
-package main
+package core
 
 import (
 	"sync"
