+++ experiment/client-go/bindings/core/wandbcore.go
@@ -1,4 +1,4 @@
-package main
+package core
 
 /*
 typedef const char cchar_t;
@@ -12,12 +12,12 @@ import "C"
 import (
 	"unsafe"
 
-	"github.com/wandb/wandb/core/internal/gowandb/internal_runopts"
 	"github.com/wandb/wandb/core/pkg/gowandb"
 	"github.com/wandb/wandb/core/pkg/gowandb/opts/runopts"
 	"github.com/wandb/wandb/core/pkg/gowandb/opts/sessionopts"
 	"github.com/wandb/wandb/core/pkg/gowandb/runconfig"
 	"github.com/wandb/wandb/core/pkg/service"
+	"github.com/wandb/wandb/gowandb/internal/gowandb/internal_runopts"
 )
 
 // globals to keep track of the wandb session and any runs
