+++ experiment/client-go/cmd/example/embed-core.bin
