+++ experiment/client-go/pkg/gowandb/session.go
@@ -4,11 +4,11 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/wandb/wandb/core/internal/execbin"
-	"github.com/wandb/wandb/core/internal/launcher"
 	"github.com/wandb/wandb/core/pkg/gowandb/opts/runopts"
 	"github.com/wandb/wandb/core/pkg/gowandb/opts/sessionopts"
 	"github.com/wandb/wandb/core/pkg/gowandb/settings"
+	"github.com/wandb/wandb/gowandb/internal/execbin"
+	"github.com/wandb/wandb/gowandb/internal/launcher"
 )
 
 type Session struct {
