+++ experiment/client-rust/src/printer.rs
